ADHITHYAA_KUMAR_S/requirements.txt
streamlit>=1.28.0
ollama>=0.1.0
flask>=2.3.0
flask-cors>=4.0.0

flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0